+++ integrations/cypress/cypress.example.env.json
@@ -2,5 +2,6 @@
   "audience": "https://thefrontside.auth0.com/api/v1/",
   "domain": "localhost:4400",
   "client_id": "00000000000000000000000000000000",
+  "connection": "Username-Password-Authentication",
   "scope": "openid profile email offline_access"
 }
