+++ integrations/cypress/cypress/tsconfig.json
@@ -1,5 +1,12 @@
 {
   "compilerOptions": {
-    "sourceMap": false
-  }
+    "target": "ESNext",
+    "lib": ["es5", "dom"],
+    "types": ["cypress", "node"],
+    "resolveJsonModule": true,
+    "moduleResolution": "node",
+    "esModuleInterop": true,
+    "noEmit": true
+  },
+  "include": ["**/*.ts"]
 }
