+++ integrations/cypress/tsconfig.json
@@ -3,7 +3,9 @@
   "compilerOptions": {
     "types": ["cypress"],
     "baseUrl": ".",
-    "noEmit": true
+    "outDir": "dist",
+    "rootDir": "src",
+    "resolveJsonModule": true
   },
   "include": [
     "src/**/*.ts",
