+++ packages/auth0/src/index.ts
@@ -50,8 +50,6 @@ export const auth0: Simulator<Options> = (slice, options) => {
 
   let handlersOptions = { ...DefaultOptions, ...options, store, services };
 
-  console.dir({ options });
-
   let auth0Handlers = createAuth0Handlers(handlersOptions);
   let openIdHandlers = createOpenIdHandlers(handlersOptions);
